requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-rest-passwordreset==1.2.1
django-taggit==2.1.0
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
